backend/requirements.txt
fastapi
uvicorn
pyserial
google-generativeai
python-dotenv
